+++ packages/xchain-thorchain-amm/__e2e__/thorchain-savers-e2e.ts
@@ -21,8 +21,8 @@ function printSaversPosition(saver: SaversPosition) {
     depositValue: saver.depositValue.formatedAssetString(),
     redeemableValue: saver.redeemableValue.formatedAssetString(),
     lastAddHeight: saver.lastAddHeight,
-    growth: saver.growth,
-    age: saver.age,
+    percentageGrowth: saver.percentageGrowth,
+    ageInDays: saver.ageInDays,
   }
   console.log(expanded)
 }