+++ packages/xchain-thorchain/src/client.ts
@@ -1,4 +1,4 @@
-import { cosmosclient, proto } from '@cosmos-client/core'
+import cosmosclient from '@cosmos-client/core'
 import {
   Balance,
   BaseXChainClient,
@@ -124,9 +124,8 @@ class Client extends BaseXChainClient implements ThorchainClient, XChainClient {
       [Network.Stagenet]: 'thorchain-stagenet-v2',
       [Network.Testnet]: 'deprecated',
     },
-    customRequestHeaders = {},
   }: XChainClientParams & ThorchainClientParams) {
-    super(Chain.Cosmos, { network, rootDerivationPaths, phrase, customRequestHeaders })
+    super(Chain.Cosmos, { network, rootDerivationPaths, phrase })
     this.clientUrl = clientUrl
     this.explorerUrls = explorerUrls
     this.chainIds = chainIds
@@ -134,15 +133,10 @@ class Client extends BaseXChainClient implements ThorchainClient, XChainClient {
     registerSendCodecs()
     registerDepositCodecs()
 
-    if (this.clientUrl[Network.Mainnet].node.includes('ninerealms.com') && !this.customRequestHeaders['x-client-id']) {
-      this.customRequestHeaders['x-client-id'] = 'xchainjs-client'
-    }
-
     this.cosmosClient = new CosmosSDKClient({
       server: this.getClientUrl().node,
       chainId: this.getChainId(network),
       prefix: getPrefix(network),
-      headers: this.customRequestHeaders,
     })
   }
 
@@ -266,7 +260,7 @@ class Client extends BaseXChainClient implements ThorchainClient, XChainClient {
    * @throws {"Phrase not set"}
    * Throws an error if phrase has not been set before
    * */
-  getPrivateKey(index = 0): proto.cosmos.crypto.secp256k1.PrivKey {
+  getPrivateKey(index = 0): cosmosclient.proto.cosmos.crypto.secp256k1.PrivKey {
     return this.cosmosClient.getPrivKeyFromMnemonic(this.phrase, this.getFullDerivationPath(index))
   }