+++ packages/xchain-thorchain/src/types/messages.ts
@@ -1,4 +1,4 @@
-import { cosmosclient, proto } from '@cosmos-client/core'
+import cosmosclient from '@cosmos-client/core'
 import { Asset } from '@xchainjs/xchain-util'
 
 export type MsgCoin = {
@@ -39,7 +39,7 @@ export type ThorchainDepositResponse = AminoWrapping<{
     memo: string
     signer: string
   }>[]
-  fee: proto.cosmos.tx.v1beta1.Fee
+  fee: cosmosclient.proto.cosmos.tx.v1beta1.Fee
   signatures: string[]
   memo: string
   timeout_height: string